% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcompendium-learning-package.R
\docType{package}
\name{Rcompendium-learning-package}
\alias{Rcompendium-learning}
\alias{Rcompendium-learning-package}
\title{Rcompendium-learning: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/samuel-a-sutherland/Rcompendium-learning}
  \item Report bugs at \url{https://github.com/samuel-a-sutherland/Rcompendium-learning/issues}
}

}
\author{
\strong{Maintainer}: Samuel Sutherland \email{Samuel.A.Sutherland@warwick.ac.uk} (\href{https://orcid.org/0000-0003-1794-9813}{ORCID}) [copyright holder]

}
\keyword{internal}
